requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
oauthlib==3.2.2
openai==1.47.0
pycharm-debug===debug
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
sniffio==1.3.1
tqdm==4.66.5
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==2.2.3
